[git-webkit] Add Publish Command https://bugs.webkit.org/show_bug.cgi?id=249586 <rdar://97397960>
Reviewed by Elliott Williams.

* Tools/Scripts/libraries/webkitscmpy/setup.py: Bump version.
* Tools/Scripts/libraries/webkitscmpy/webkitscmpy/__init__.py: Ditto.
* Tools/Scripts/libraries/webkitscmpy/webkitscmpy/mocks/local/git.py: Add 'git branch -a --merged'
* Tools/Scripts/libraries/webkitscmpy/webkitscmpy/program/__init__.py:
* Tools/Scripts/libraries/webkitscmpy/webkitscmpy/program/publish.py: Added.
(Publish):
(Publish.parser): User provides command a list of refs to publish.
(Publish.branches_on): List all branches that are on the history of a specified ref.
(Publish.tags_on): List all tags that are on the history of a specified ref.
(Publish.parental_intersection): Find the intersection point of the provided commit and the parent
branch of that commit. Note that the provided commit might be on the parent branch, in which case
this function returns nothing.
(Publish._push_branch_ref): Push a branch ref onto the provided mapping. This function takes into
consideration any existing branch ref of the same name, taking the more up to date of the two.
(Publish.main): Given the refs provided by the user, find all refs on that history and push those refs
to the "next" source remote.
* Tools/Scripts/libraries/webkitscmpy/webkitscmpy/program/pull_request.py:
(PullRequest.create_pull_request): Prefer the most secret remote.
* Tools/Scripts/libraries/webkitscmpy/webkitscmpy/test/publish_unittest.py: Added.
(TestPublish): Added tests.
* metadata/git_config_extension: Sort source remotes.

Canonical link: https://commits.webkit.org/260044@main