Cocoa availability macros are not working for non-WebKit sources on buildbots https://bugs.webkit.org/show_bug.cgi?id=251989 rdar://105221617
Reviewed by NOBODY (OOPS!).

For public SDK use-cases, such as WebKit iOS, TvOS, WatchOS build bots
and non-Apple contributors:

configure-xcode-for-embedded-development would remove the
availability macros for any code that is compiled with the TvOS,
WatchOS, iOS SDKs.

This causes compile errors in comprehensive switch statements, as
removing the availability may add new enumeration cases that are not
valid for the platform.

Using the invalid enumeration cases is problematic for non-WebKit
codebases such as ANGLE that uses Metal API. Since the upstream likely
does not have the modified SDK, the invalid enumeration cases will cause
errors for the upstream compile.

Solve this by moving the availability macro erasure to affect all
WTF projects, not all projects compiled with WebKit SDK.

The erasure is invoked from Platform.h include chain.
The erasure is not invoked from config.h of the various projects since
since config.h and *Prefix.h files both include platform headers that
already need the modification. If it was applied
only to various config.h files, then the *Prefix.h files would cause
platform includes that already would need the override.

The most beneficial would be to apply the erasure per-compilation unit
basis. This is not done because of the above problem. Also, the
compilation units would need to not be built with unified build as they
are today. Changing them to @no-unify surfaces compilation logic errors
that are currently hidden due to unified build. Example of how the
per-compilation unit erasure would work is shown with the modifications
to bmalloc AvailableMemory.cpp.

* Source/WTF/WTF.xcodeproj/project.pbxproj:
* Source/WTF/wtf/CMakeLists.txt:
* Source/WTF/wtf/Platform.h:
* Source/WTF/wtf/PlatformMacroOverrides.h: Added.
* Source/bmalloc/PlatformMac.cmake:
* Source/bmalloc/bmalloc.xcodeproj/project.pbxproj:
* Source/bmalloc/bmalloc/AvailableMemory.cpp:
(bmalloc::memorySizeAccordingToKernel):
* Source/bmalloc/bmalloc/darwin/BAllowProhibitedAPIs.h: Added.
* Tools/Scripts/configure-xcode-for-embedded-development: