[GStreamer][VideoCapture] Prefer raw formats over encoded formats https://bugs.webkit.org/show_bug.cgi?id=252004
Reviewed by NOBODY (OOPS!).

The problem with a video capturer emitting encoded formats is that they need re-encoding when the
format doesn't match the selected outgoing RTP video capabilities. So for instance if a jpeg format
is selected and we have no hardware decoder for it, we have a new CPU bottleneck in order to emit
the video stream to a remote peer expecting a different format.

* Source/WebCore/platform/mediastream/gstreamer/GStreamerVideoCapturer.cpp:
(WebCore::DeviceCapsProcessHolder::DeviceCapsProcessHolder):
(WebCore::GStreamerVideoCapturer::reconfigure):