Increase underline intervals when using `text-decoration-skip-ink` https://bugs.webkit.org/show_bug.cgi?id=252042
Reviewed by NOBODY (OOPS!).

When using `text-decoration-skip-ink`, intervals for skipped descenders
or glyphs are represented as a `DashArray` where the first float is
the beginning of the spacing and the second is the end.

When displayed, these values will be rounded to integers, resulting in
an inconsistently shorter or longer interval.

To solve it, we now always `floor` the begining of the skipped interval
and `ceiling` the end.

Fixes `fast/css3-text/css3-text-decoration/text-decoration-skip/text-decoration-skip-ink.html`.

* Source/WebCore/platform/graphics/FontCascade.cpp:
(WebCore::FontCascade::dashesForIntersectionsWithRect const):