Implement margin-trim for legacy inline layout https://bugs.webkit.org/show_bug.cgi?id=252062 rdar://problem/105283580
Reviewed by NOBODY (OOPS!).

When floats are being placed, we can check to see if we can trim some
of its margins based on its surrounding constraints. In particular, we
can check to see if there are any other items on the left/right
(depending on the position of the float) of the box at that candidate
vertical position. If there is nothing there (i.e. the float would be
placed flush against its containing block), then we can trim the
respective margin. However, if there is an item at that vertical, then
we must wait until the float gets its final position before tirmming
any of its margins. This is because trimming a margin too early could
result it in being placed in a position where it should not be.

<div style="margin-trim: inline;">
  <div style="float: left; width: 50px; height: 50px; background-color: green; margin-inline-start: 30px;"></div>
  <div style="float: right; width: 50px; height: 50px; background-color: red; margin-inline-end: 30px;"></div>
</div>
In this simple example we can easily trim both the margins of the floats
as we attempt to place them. Inside BreakingContext::handleFloat() we
check the logical offset for the line against the logical offset for
content inside the containing block. If these two values are equal,
then the float would be placed against the box and we can trim the
margin early.

In all cases this may not entirely be possible:
<style>
container {
    display: block;
    width: 110px;
    margin-trim: inline;
    font-size: 0px;
}
item {
    display: block;
    width: 100px;
    height: 25px;
    background-color: green;
}
.inline-block {
    display: inline-block;
    height: 50px;
}
.float-left {
    float: left;
}
.overflowing {
    margin-inline-start: 55px;
}
</style>
<container>
    <item class="inline-block"></item>
    <item class="float-left"></item>
    <item class="float-left overflowing"></item>
</container>
For the last float we see that there is content on the left side at
this vertical position, so we do not trim the margin here. Instead we
have to wait until this float gets pushed under all the other ones
before trimming.

If we trim too early, the float could get positioned in a spot where it
would not normally fit with its margin. For this we fall back to the
code in RenderBlockFlow::computeLogicalLocationForFloat that will trim
the margin at its final position. This is the same logic that was used
to trim the margins for floats in block containers that did not have
any inline children.

* LayoutTests/fast/inline/legacy-margin-trim/block-container-float-block-end-inline-layout-expected.html: Added
* LayoutTests/fast/inline/legacy-margin-trim/block-container-float-block-end-inline-layout.html: Added
* LayoutTests/fast/inline/legacy-margin-trim/block-container-float-block-start-inline-layout-expected.html: Added
* LayoutTests/fast/inline/legacy-margin-trim/block-container-float-block-start-inline-layout.html: Added
* LayoutTests/fast/inline/legacy-margin-trim/block-container-float-inline-end-inline-layout-expected.html: Added
* LayoutTests/fast/inline/legacy-margin-trim/block-container-float-inline-end-inline-layout.html: Added
* LayoutTests/fast/inline/legacy-margin-trim/block-container-float-inline-inline-layout-expected.html: Added
* LayoutTests/fast/inline/legacy-margin-trim/block-container-float-inline-inline-layout.html: Added
* LayoutTests/fast/inline/legacy-margin-trim/block-container-float-inline-start-inline-layout-expected.html: Added
* LayoutTests/fast/inline/legacy-margin-trim/block-container-float-inline-start-inline-layout.html: Added
* LayoutTests/fast/inline/legacy-margin-trim/block-container-float-min-content-inline-layout-expected.html: Added
* LayoutTests/fast/inline/legacy-margin-trim/block-container-float-min-content-inline-layout.html: Added
* LayoutTests/fast/inline/legacy-margin-trim/block-container-float-multiple-formatting-contexts-inline-layout-expected.html: Added
* LayoutTests/fast/inline/legacy-margin-trim/block-container-float-multiple-formatting-contexts-inline-layout.html: Added
* LayoutTests/fast/inline/legacy-margin-trim/block-container-float-overflowing-margin-has-margin-trimmed-at-final-position-expected.html: Added
* LayoutTests/fast/inline/legacy-margin-trim/block-container-float-overflowing-margin-has-margin-trimmed-at-final-position.html: Added
* LayoutTests/fast/inline/legacy-margin-trim/block-container-float-trimmed-margin-allows-float-to-fit-inline-layout-expected.html: Added
* LayoutTests/fast/inline/legacy-margin-trim/block-container-float-trimmed-margin-allows-float-to-fit-inline-layout.html: Added
Source/WebCore/layout/integration/LayoutIntegrationCoverage.cpp:
(WebCore::LayoutIntegration::canUseForChild):
* Source/WebCore/rendering/LegacyLineLayout.cpp:
(WebCore::LegacyLineLayout::contentBoxLogicalHeight const):
* Source/WebCore/rendering/LegacyLineLayout.h:
* Source/WebCore/rendering/RenderBlockFlow.cpp:
(WebCore::RenderBlockFlow::layoutBlock):
(WebCore::RenderBlockFlow::blockFormattingContextInFlowContentHeight const):
(WebCore::RenderBlockFlow::blockFormattingContextInFlowBlockLevelContentHeight const): Deleted.
* Source/WebCore/rendering/RenderBlockFlow.h:
* Source/WebCore/rendering/line/BreakingContext.h:
(WebCore::BreakingContext::handleFloat):