Make WEBKIT_LIBRARIES directory in built-product-archive https://bugs.webkit.org/show_bug.cgi?id=252085
Reviewed by Ryan Haddad.

The WinCairo bots download their requirement libraries to
`WebKitLibraries/win`, however in 260060@main the
`WebKitLibraries/win/tools` directory was removed so a git checkout no
longer contained the directory. The `built-product-archive` script tries
to put metadata for what requirements to download in that directory but
it started failing when bots restarted because that directory was no
longer there.

Check that the environment variable `WEBKIT_LIBRARIES` is set and then
use that to create the directory if necessary.

* Tools/CISupport/built-product-archive:

Canonical link: https://commits.webkit.org/260131@main