Unreviewed build fix for Clang-libstdc++ combo https://bugs.webkit.org/show_bug.cgi?id=252156
Unreviewed build fix after 260175@main for Clang and libstdc++ builds.

Assignment of CSSRadialGradientValue::GradientBox variant with a
CSSGradientPosition value is causing compilation failures due to other variant
types that are somehow treated as constructible in this case, due to
CSSGradientPosition being a pair of Ref<CSSValue> values and other variants
having a pair of Ref<CSSPrimitiveValue> values as their first members.

This is avoidable by reconstructing the variant and using the std::in_place_type
helper to explicitly specify that the CSSGradientPosition value should be used
in the variant.

* Source/WebCore/css/parser/CSSPropertyParserHelpers.cpp:
(WebCore::CSSPropertyParserHelpers::consumeRadialGradient):

Canonical link: https://commits.webkit.org/260186@main