[WGSL] Implement relational expression parsing https://bugs.webkit.org/show_bug.cgi?id=252152 rdar://problem/105379078
Reviewed by Tadeu Zagallo.

Handle parsing of relational expressions according to WGSL spec. This introduces
the need to account for the ambiguity when trying to distinguish the
use of < as either the less-than operator or the beginning of a
type-parameter list. The solution chosen here is to check for one of the vector
or matrix type names before consuming `<` to produce a callable type constructor.

Alternatively, those names could be turned into keywords to aid the parser, or
we could save the lexer state post identifier and back track on failure to parse
a type constructor.

Canonical link: https://commits.webkit.org/260189@main