Implement parser for text-spacing and its longhands for auto and none https://bugs.webkit.org/show_bug.cgi?id=252068 rdar://105094870 Reviewed by NOBODY (OOPS!).
Reference: https://github.com/w3c/csswg-drafts/issues/4246#issuecomment-1404738513

Implement a parser with longhands support for:
text-spacing: auto | none | <text-autospace> || <text-spacing-trim>

text-autospace: auto | no-autospace
text-spacing-trim: auto | space-all

We can then iterate from here adding support for 'normal' in the
shorthand and for the remaining values in the longhands.

For the text-spacing shorthand, apart from appearing as a single token,
'auto' can also appear as belonging to one of the longhands in any order.
In case we parse an 'auto' value and the token range is not exhauted,
there is a certain ambiguity if it belongs to the first or second
longhand, since they can come in any order.
Becuse all the other values of the longhands are unique, we can peek
into the next token to solve this ambiguity.

For the common case, where 'auto' doesn't appear as the first token,
we try to re-use the parsing functions for the longhands. For that, we
need to divide the range into 2 subranges, one for each longhand.
We use again our knowledge of unique values for a longhand to identify
such subranges.

* LayoutTests/TestExpectations:
* Source/WebCore/animation/CSSPropertyAnimation.cpp:
(WebCore::CSSPropertyAnimationWrapperMap::CSSPropertyAnimationWrapperMap):
* Source/WebCore/css/CSSProperties.json:
* Source/WebCore/css/ComputedStyleExtractor.cpp:
(WebCore::textSpacingFromStyle):
(WebCore::ComputedStyleExtractor::valueForPropertyInStyle):
* Source/WebCore/css/ShorthandSerializer.cpp:
(WebCore::ShorthandSerializer::serialize):
(WebCore::ShorthandSerializer::serializeTextSpacing const):
* Source/WebCore/css/parser/CSSPropertyParser.cpp:
(WebCore::CSSPropertyParser::consumeTextSpacing):
(WebCore::CSSPropertyParser::parseShorthand):
* Source/WebCore/css/parser/CSSPropertyParser.h:
* Source/WebCore/css/parser/CSSPropertyParserHelpers.h:
(WebCore::CSSPropertyParserHelpers::isValueIDUniqueToTextSpacingTrimLonghand):
(WebCore::CSSPropertyParserHelpers::isValueIDUniqueToTextAutospaceLonghand):
* Source/WebCore/platform/text/TextSpacing.h:
(WebCore::operator<<):
* Source/WebCore/rendering/style/RenderStyle.h: