[browserperfdash-benchmark] Add optional --timestamp parameter to send the data to the server with a custom date. https://bugs.webkit.org/show_bug.cgi?id=252355
Reviewed by Dewei Zhu.

Add an optional paramter '--timestamp' that allows to send the data
to the server marked with a specific date. So then the dashboard
will add the benchmark data as it was generated on the date specified.

This is useful for the following two use cases: 1) Several machines
downloading builds and doing performance tests in parallel are not
guaranteed to end the testing in the order they downloaded the
builds. So by passing a timestamp that can be obtained from the git
revision of the build (or even from the build executable timestamp)
the data gets sorted properly on the dashboard.
2) Some data for older versions is missed, this way it is possible
to tell the bot to run the benchmarks with an older version and then
have the data ordered properly on the dashboard.

* Tools/Scripts/webkitpy/browserperfdash/browserperfdash_runner.py:
(parse_args):
(BrowserPerfDashRunner.__init__):

Canonical link: https://commits.webkit.org/260345@main